Fix admin console + public dashboard correctness — audit batch 16: AUDIT-185, 229, 230, 232, 233, 234, 235 (v0.11.224)
Seven frontend/handler findings from the 2026-08-27 engineering audit, each with a
content-pinning regression test in internal/shell/ that fails on revert.

- AUDIT-185: device-detail public-interface toggle. loadPublicInterfaces() now
  returns its fetch so the init chain waits, and a publicIfacesLoaded flag blocks
  the toggle / disables the checkboxes until the map loads — a pre-load click could
  POST an empty map and wipe every other device's public-dashboard selection.
- AUDIT-229: removed the dead admin.GET("/network") route (network.html,
  admin-network.js and the sidebar link were already deleted; the route errored).
- AUDIT-230: connection-detail tunnel charts preserve the operator's
  Combined/Transfer mode across the 30s auto-refresh. The innerHTML rebuild
  destroyed the .fwmon-bw-toggle mount() reads; the mode is now snapshotted per
  group before the wipe and passed to mount as initialView.
- AUDIT-232: a fresh IPSec deploy/rollback/recheck no longer flashes the prior
  deploy's failure banner or double-polls. openDeployModal takes a deferPoll flag;
  a launch defers the single poll until its POST is accepted.
- AUDIT-234: rich VPN detail panel rendered dead for any device with a live tunnel
  — bare AC.formatTunnelUptime threw ReferenceError inside .map(). AC hoisted to
  module scope in diagram-panels.js.
- AUDIT-233: shared root cause with 234 in admin-threatintel.js. Corrected scope:
  the ReferenceError was CAUGHT by the chained .catch, so it was never an unhandled
  rejection and the button was never stuck; the only real effects were no success
  toast and onFeedToggle's table not refreshing after a successful toggle. Module-
  scope AC (the shared fix) restores both.
- AUDIT-235: public dashboard 15m/30m ranges, fixed end to end. Client now uses
  parseFloat (parseInt truncated 0.25/0.5 to 0); GetPublicInterfaceChart parses the
  range as a float and drives a sub-hour cutoff (was Atoi -> 1h), and
  GetPublicStatusHistory honors a sub-hour hours value via an explicit cutoff (was
  the integer ParseHours 24h default). Integer ranges keep the shared helpers.

Bumps ServerVersion to 0.11.224.

+++ cmd/api/static/js/public-dashboard.js
@@ -626,7 +626,7 @@
             bwView = e.target.value;
             refreshBandwidthCharts();
         } else if (e.target.id === 'dash-range') {
-            dashRange = parseInt(e.target.value);
+            dashRange = parseFloat(e.target.value); // AUDIT-235: parseInt truncated 0.25/0.5 (15m/30m) to 0
             refreshBandwidthCharts();
             refreshCpuCharts();
         }
